+++ lib/dbf/memo/dbase3.rb
@@ -7,7 +7,12 @@ module DBF
         data.seek offset(start_block)
         memo_string = +''
         loop do
-          block = data.read(BLOCK_SIZE).gsub(/(\000|\032)/, '')
+          block = data.read(BLOCK_SIZE)
+          # A start block past EOF yields nil; return what we have rather
+          # than crashing on nil.gsub.
+          break if block.nil?
+
+          block = block.gsub(/(\000|\032)/, '')
           memo_string << block
           break if block.size < BLOCK_SIZE
         end